pyproject.toml
[tool.poetry]
name = "mfgfn"
version = "0.2.0"
description = "Multi-fidelity active learning with GFlowNets"
authors = ["Nikita <nikitasaxena0209@gmail.com> and Alex <alex.hernandez-garcia@mila.quebec>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
# Base dependencies.
botorch = ">=0.8.2"
gflownet = { git = "https://github.com/alexhernandezgarcia/gflownet", branch = "cont_mf" }
hydra-core = ">=1.3.1"
matplotlib = "*"
numpy = "*"
pandas = "*"
python = "^3.10"
rdkit = ">=2022.9.5"
selfies = ">=2.1.1"
six = ">=1.16.0"
torch = { version = "==1.13.1", source = "torch-gpu" }
torchtyping = ">=0.1.4"
tqdm = "*"
wandb = ">=0.13.9"

# Optional dependencies (extras)

# Dev extras
black = { version = ">=22.12.0", optional = true }
flake8 = { version = ">=6.1.0", optional = true }
isort = { version = ">=5.12.0", optional = true }
pylint = { version = ">=2.17.6", optional = true }
ipdb = { version = ">=0.13.11", optional = true }

[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "primary"

[[tool.poetry.source]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu117"
priority = "primary"

[tool.poetry.extras]
dev = [
    "black",
    "flake8",
    "isort",
    "pylint",
    "ipdb",
]
all = [
    "black",
    "flake8",
    "isort",
    "pylint",
    "ipdb",
]
